requirements.txt
Flask>=2.0.2
Flask-SQLAlchemy>=3.0.2
bcrypt==4.0.1
flask-sock==0.6.0
Werkzeug==2.2.3
mysql-connector-python
